// src/shared/types/errors.ts
export enum ErrorCode {
  EXTRACTION_FAILED = 'EXTRACTION_FAILED',
  AUTH_EXPIRED = 'AUTH_EXPIRED',
  AUTH_DEVICE_FLOW_DENIED = 'AUTH_DEVICE_FLOW_DENIED',
  AUTH_DEVICE_FLOW_EXPIRED = 'AUTH_DEVICE_FLOW_EXPIRED',
  GITHUB_RATE_LIMITED = 'GITHUB_RATE_LIMITED',
  GITHUB_CONFLICT = 'GITHUB_CONFLICT',
  GITHUB_PERMISSION_DENIED = 'GITHUB_PERMISSION_DENIED',
  GITHUB_REPO_NOT_FOUND = 'GITHUB_REPO_NOT_FOUND',
  AI_GENERATION_FAILED = 'AI_GENERATION_FAILED',
  AI_SCHEMA_VALIDATION_FAILED = 'AI_SCHEMA_VALIDATION_FAILED',
  AI_KEY_INVALID = 'AI_KEY_INVALID',
  NETWORK_TIMEOUT = 'NETWORK_TIMEOUT',
  QUEUE_OVERFLOW = 'QUEUE_OVERFLOW',
  IDEMPOTENCY_COLLISION = 'IDEMPOTENCY_COLLISION',
}